engine/classification/requirements.txt
# requirements
tqdm
numpy
scikit-learn==1.1.3
Pillow==9.3.0
opencv-python==4.6.0.66
matplotlib==3.6.2
timm==0.6.11
tensorboardX
